afstand-->int

+++ Queries voor db.sql
@@ -37,7 +37,7 @@ CREATE TABLE route
     PersonID int NULL,
     AantalPakketten int not null,
     Reistijd decimal(4,2),
-    Afstand decimal(5,2) not null,
+    Afstand int(11) not null,
     Provincie varchar(45) not null,
     Status varchar(50) Not null,
     Opmerkingen varchar(200) null,
